moving methods that deal with using inventory items to specific method modules from the game controller: gameMethods.useInvItem, gameMethods.hurtCharacter, gameMethods.endOfGame,  roomMethods.check4Monster, charMethods.deleteFromInventory, charMethods.batt leMonster - these interact across the modules
