+++ src/models/character.js
@@ -2,6 +2,7 @@ module.exports = {
 
   name: 'Adventurer',
   hp: 100,
-  inventory: []
+  inventory: [],
+  isDead: false
 
 };
